fix(broadcast): move isSafeBannerLink out of schemas.ts to keep it out of the client bundle
